Create Monte Carlo simulation to solve for V3 with 100 sets of R's
A normal resistor has 5% tolerance with a uniform distribution: R = (1 + 0.05x)R0 where x is uniform(-1, 1). This code allows me to find the expected variation in V3 with a 5% tolerance for each resistor in the circuit. When designing a circuit to be manufactured this code would allow you to adjust your code in accordance with the expected variation in V3.